Update SelfieListActivity.java
The line that I added will force a front facing camera if available.  Just a nice touch for a "Selfie" app. I am working on this too for Coursera I did mine a little different.  I just chucked photos in a folder and loop through the files in the folder with png extensions and deal with them that way.  You have a much more complete solution though. Good Luck!